requirements.txt
tensorflow==2.14.0
numpy
pillow
matplotlib
gitpython
